Move VMC code to Sampler/ directory
